migrate codebase from section2 to section3 and move user credentials from config to in-memory UserDetailsService bean
- Renamed package structure from springsecsection2 to springsecsection3
- Updated application main class, configs, and all controllers accordingly
- Renamed test class to match new section
- Modified application.yml to reflect user config changes
- Updated pom.xml for necessary adjustments
- Removed SPRING_SECURITY_USER_NAME and SPRING_SECURITY_USER_PASSWORD from application.yml
- Introduced an in-memory UserDetailsService bean with static users 'user' and 'admin'
- Preserved application name and logging configurations
