+++ section3/src/main/java/com/eazybytes/springsecsection3/EazyBankBackendApplication.java
@@ -1,4 +1,4 @@
-package com.eazybytes.springsecsection1;
+package com.eazybytes.springsecsection3;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
