+++ section3/src/main/java/com/eazybytes/springsecsection3/configs/ProjectSecurityConfig.java
@@ -1,10 +1,14 @@
-package com.eazybytes.springsecsection1.configs;
+package com.eazybytes.springsecsection3.configs;
 
 import org.springframework.boot.autoconfigure.security.SecurityProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.core.annotation.Order;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.core.userdetails.User;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.provisioning.InMemoryUserDetailsManager;
 import org.springframework.security.web.SecurityFilterChain;
 
 import static org.springframework.security.config.Customizer.withDefaults;
@@ -27,5 +31,22 @@ public class ProjectSecurityConfig {
         http.formLogin(withDefaults());
         http.httpBasic(withDefaults());
         return http.build();
+    }
+
+    @Bean
+    public UserDetailsService userDetailsService() {
+        UserDetails user =
+                User.withUsername("user")
+                        .password("1234")
+                        .authorities("read")
+                        .build();
+
+        UserDetails admin =
+                User.withUsername("admin")
+                        .password("1234")
+                        .authorities("admin")
+                        .build();
+
+        return new InMemoryUserDetailsManager(user, admin);
     }
 }
