+++ section3/src/main/java/com/eazybytes/springsecsection3/controllers/ContactController.java
@@ -1,4 +1,4 @@
-package com.eazybytes.springsecsection1.controllers;
+package com.eazybytes.springsecsection3.controllers;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
