+++ section3/src/main/resources/application.yml
@@ -1,10 +1,6 @@
 spring:
   application:
     name: ${SPRING_APP_NAME:spring-security-sec1}
-  security:
-    user:
-      name: ${SPRING_SECURITY_USER_NAME:admin}
-      password: ${SPRING_SECURITY_USER_PASSWORD:admin}
 logging:
   pattern:
     console: ${LOGPATTERN_CONSOLE:%green(%d{HH:mm:ss.SSS}) %blue(%-5level) %red([%thread]) %yellow(%logger{15}) - %msg%n}
