+++ section3/src/test/java/com/eazybytes/springsecsection3/EazyBankBackendApplicationTests.java
@@ -1,4 +1,4 @@
-package com.eazybytes.springsecsection1;
+package com.eazybytes.springsecsection3;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
